+++ src/main/java/com/flowiee/dms/base/BaseCurdService.java
@@ -3,11 +3,11 @@ package com.flowiee.dms.base;
 import java.util.Optional;
 
 public interface BaseCurdService<T> {
-    Optional<T> findById(Integer entityId);
+    Optional<T> findById(Long entityId);
 
     T save(T entity);
 
-    T update(T entity, Integer entityId);
+    T update(T entity, Long entityId);
 
-    String delete(Integer entityId);
+    String delete(Long entityId);
 }
